reqif: switch to selecting parser via an argument
ReqIF structure seems to be too different for various documents.
Let each document type have its own dedicated ReqIF parser which also
simplified each parser's structure.
