+++ .pylint.ini
@@ -6,7 +6,7 @@
 extension-pkg-whitelist=
 
 # Specify a score threshold to be exceeded before program exits with error.
-fail-under=9.0
+fail-under=9.1
 
 # Add files or directories to the blacklist. They should be base names, not
 # paths.
