+++ tests/integration/reqif/91-example-magna/test.itest
@@ -1,3 +1,5 @@
+UNSUPPORTED: true
+
 RUN: %strictdoc import reqif-experimental %S/sample.reqif %S/output/output.sdoc --mapping=magna
 RUN: %strictdoc export %S/output/output.sdoc
